Лабораторная 11. 2 подход, исправленно

+++ unitl11.h
@@ -1,4 +1,4 @@
 #ifndef unitl11_h
 #define unitl11_h
-void perem_arg(int k, int *sum, int *max, int *min, int *sred, ...);
+void perem_arg(int k, int *sum, int *max, int *min, double *sred, ...);
 #endif
